requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
PyJWT==1.7.1
sqlparse==0.5.0
typing-extensions==4.11.0
tzdata==2024.1
